For 1.7-, removing enumerator alpha assertion.
Maps in go don't guarentee the order of the contents the return.
Starting in go 1.8, there are premade sort methods in the standard
library. With that addition, it makes it more worthwhile to guarentee
that the Trie returns the strings in alphabetical order. However, given
the low pri nature of it, it's not worth back porting some sorting
solution for older Go releases.